Moved fixture items to configuration file
